Refactorización de Personaje
Se refactorizaron las clases/interfaces de Repository, Service, ServiceImpl, y Controller de Personaje
