+++ DisneyApi/src/main/java/com/project/DisneyApi/repository/PersonajeRepository.java
@@ -3,13 +3,12 @@ package com.project.DisneyApi.repository;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.project.DisneyApi.entity.Personaje;
 
 @Repository
-public interface PersonajeRepository extends JpaRepository<Personaje, Integer>{
+public interface PersonajeRepository extends BaseRepository<Personaje, Long>{
 	
 	Optional<Personaje>findByNombre(String nombre);
 	boolean existsByNombre(String nombre);
